feat(): cleanup test AFSs too

+++ test/destroy.js
@@ -33,13 +33,14 @@ test.beforeEach(async (t) => {
   } catch (err) {
     console.log(err)
   }
-  t.context = { afs, idPath: aid.createIdentityKeyPath(afs.ddo) }
+  t.context = { afs, idPath: aid.createIdentityKeyPath(afs.ddo), afsPath: createAFSKeyPath(afs.did) }
 })
 
 test.afterEach(async (t) => {
-  const { idPath } = t.context
-  if (idPath) {
+  const { idPath, afsPath } = t.context
+  if (idPath && afsPath) {
     await pify(rimraf)(idPath)
+    await pify(rimraf)(afsPath)
   }
 })
